+++ server/src/modules/user.module.js
@@ -40,6 +40,7 @@ const userSchema = new mongoose.Schema(
       type: String,
       default: "",
     },
+    groups: [{ type: mongoose.Schema.Types.ObjectId, ref: "Group" }],
   },
   { timestamps: true }
 );